safari-600.7-branch/LayoutTests/crypto/subtle/postMessage-worker.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<script src="../../resources/js-test-pre.js"></script>
<script src="resources/common.js"></script>
</head>
<body>
<script>

description("Test sending crypto keys via postMessage to a worker.");

jsTestIsAsync = true;

importTestKeys().then(function(keys) {
    var worker = new Worker("resources/postMessage-worker.js");
    worker.onmessage = function(evt) {
        if (!evt.data.result) {
            testFailed("Check failed in worker: " + evt.data.message);
        } else {
            testPassed("All checks passed in worker");
            key = evt.data.key;
            shouldBe("key.type", "'secret'");
            shouldBe("key.extractable", "true");
            shouldBe("key.algorithm.name", "'HMAC'");
            shouldBe("key.algorithm.length", "16");
            shouldBe("key.usages", '["decrypt", "encrypt", "sign", "verify"]');
        }
        finishJSTest();
    }
    worker.postMessage(keys.hmacSha1);
});
</script>
<script src="../../resources/js-test-post.js"></script>
</body>
</html>
